+++ NEWS.md
@@ -1,5 +1,12 @@
 # User visible changes in `FITSBase` package
 
+
+## Version 0.2.1
+
+- Can search cards in FITS header by regular expressions.
+- `length(eachmatch(pat,hdr::FitsHeader))` yields number of matches.
+- Some bug fixes.
+
 ## Version 0.2.0
 
 - Package renamed `FITSBase.jl`.